// src/common/enums/report-status.enum.ts
export enum ReportStatus {
  READY_TO_GENERATE = 'READY_TO_GENERATE',
  GENERATING_REPORT = 'GENERATING_REPORT',
  INFORMATION_PENDING = 'INFORMATION_PENDING',
  NEEDS_ANALYSIS = 'NEEDS_ANALYSIS',
  ERROR_PROCESSING = 'ERROR_PROCESSING',
  READY_TO_SEND = 'READY_TO_SEND',
  SENT = 'SENT',
  ERROR_SEND = 'ERROR_SEND',
}